+++ main.db.go
@@ -7,7 +7,6 @@ import (
 
 	//SQL Package
 
-	hornbillHelpers "github.com/hornbill/goHornbillHelpers"
 	"github.com/jmoiron/sqlx"
 
 	//SQL Drivers
@@ -22,13 +21,13 @@ func buildConnectionString() string {
 	if importConf.DBConf.Database == "" ||
 		importConf.DBConf.Authentication == "SQL" && (importConf.DBConf.UserName == "" || importConf.DBConf.Password == "") {
 		//Conf not set - log error and return empty string
-		hornbillHelpers.Logger(4, "Database configuration not set.", true, logFileName)
+		logger(4, "Database configuration not set.", true, true)
 		return ""
 	}
 	if importConf.DBConf.Driver != "odbc" {
-		hornbillHelpers.Logger(1, "Connecting to Database Server: "+importConf.DBConf.Server, true, logFileName)
+		logger(1, "Connecting to Database Server: "+importConf.DBConf.Server, true, true)
 	} else {
-		hornbillHelpers.Logger(1, "Connecting to ODBC Data Source: "+importConf.DBConf.Database, true, logFileName)
+		logger(1, "Connecting to ODBC Data Source: "+importConf.DBConf.Database, true, true)
 	}
 
 	connectString := ""
@@ -77,37 +76,37 @@ func buildConnectionString() string {
 func queryDatabase(delete bool) error {
 	connString := buildConnectionString()
 	if connString == "" {
-		hornbillHelpers.Logger(4, " [DATABASE] Database Connection String Empty. Check the DBConf section of your configuration.", true, logFileName)
+		logger(4, " [DATABASE] Database Connection String Empty. Check the DBConf section of your configuration.", true, true)
 		return errors.New("database connection string empty - check the dbconf section of your configuration")
 	}
 	//Connect to the JSON specified DB
 	db, err := sqlx.Open(importConf.DBConf.Driver, connString)
 	if err != nil {
-		hornbillHelpers.Logger(4, " [DATABASE] Database Connection Error: "+fmt.Sprintf("%v", err), true, logFileName)
+		logger(4, " [DATABASE] Database Connection Error: "+fmt.Sprintf("%v", err), true, true)
 		return err
 	}
 	defer db.Close()
 	//Check connection is open
 	err = db.Ping()
 	if err != nil {
-		hornbillHelpers.Logger(4, " [DATABASE] [PING] Database Ping Error: "+fmt.Sprintf("%v", err), true, logFileName)
+		logger(4, " [DATABASE] [PING] Database Ping Error: "+fmt.Sprintf("%v", err), true, true)
 		return err
 	}
-	hornbillHelpers.Logger(3, "[DATABASE] Connection Successful", true, logFileName)
+	logger(3, "[DATABASE] Connection Successful", true, true)
 	sqlQuery := importConf.Query
 	if delete {
-		hornbillHelpers.Logger(3, "[DATABASE] Running database query for asset relationship removals. Please wait...", true, logFileName)
+		logger(3, "[DATABASE] Running database query for asset relationship removals. Please wait...", true, true)
 		sqlQuery = importConf.RemoveQuery
 	} else {
-		hornbillHelpers.Logger(3, "[DATABASE] Running database query for asset relationships. Please wait...", true, logFileName)
+		logger(3, "[DATABASE] Running database query for asset relationships. Please wait...", true, true)
 
 	}
 
-	hornbillHelpers.Logger(3, "[DATABASE] Query: "+sqlQuery, false, logFileName)
+	logger(3, "[DATABASE] Query: "+sqlQuery, false, true)
 	//Run Query
 	rows, err := db.Queryx(sqlQuery)
 	if err != nil {
-		hornbillHelpers.Logger(4, " [DATABASE] Database Query Error: "+fmt.Sprintf("%v", err), true, logFileName)
+		logger(4, " [DATABASE] Database Query Error: "+fmt.Sprintf("%v", err), true, true)
 		return err
 	}
 	defer rows.Close()
@@ -120,7 +119,7 @@ func queryDatabase(delete bool) error {
 		results := make(map[string]interface{})
 		err = rows.MapScan(results)
 		if err != nil {
-			hornbillHelpers.Logger(4, " [DATABASE] Data Unmarshal Error: "+fmt.Sprintf("%v", err), true, logFileName)
+			logger(4, " [DATABASE] Data Unmarshal Error: "+fmt.Sprintf("%v", err), true, true)
 		} else {
 			//Stick marshalled data map in to parent slice
 			if delete {
@@ -132,9 +131,9 @@ func queryDatabase(delete bool) error {
 		}
 	}
 	if delete {
-		hornbillHelpers.Logger(3, "[DATABASE] "+strconv.Itoa(intAssetSuccess)+" of "+strconv.Itoa(intAssetCount)+" asset relationship removal records successfully retrieved ready for processing.", true, logFileName)
+		logger(3, "[DATABASE] "+strconv.Itoa(intAssetSuccess)+" of "+strconv.Itoa(intAssetCount)+" asset relationship removal records successfully retrieved ready for processing.", true, true)
 	} else {
-		hornbillHelpers.Logger(3, "[DATABASE] "+strconv.Itoa(intAssetSuccess)+" of "+strconv.Itoa(intAssetCount)+" asset relationship records successfully retrieved ready for processing.", true, logFileName)
+		logger(3, "[DATABASE] "+strconv.Itoa(intAssetSuccess)+" of "+strconv.Itoa(intAssetCount)+" asset relationship records successfully retrieved ready for processing.", true, true)
 	}
 	return nil
 }