+++ main.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"encoding/json"
-	"encoding/xml"
-	"errors"
 	"flag"
 	"fmt"
 	"os"
@@ -12,12 +10,12 @@ import (
 
 	apiLib "github.com/hornbill/goApiLib"
 	hornbillHelpers "github.com/hornbill/goHornbillHelpers"
+	"github.com/tcnksm/go-latest"
 )
 
 func main() {
 	var err error
-	//-- Start Time for Duration
-	startTime = time.Now()
+
 	//-- Start Time for Log File
 	timeNow = time.Now().Format("20060102150405")
 	logFileName = "assetRelationships" + timeNow + ".log"
@@ -40,20 +38,10 @@ func main() {
 	//Create shared espxmlmc session
 	espXmlmc = apiLib.NewXmlmcInstance(importConf.InstanceID)
 	espXmlmc.SetAPIKey(importConf.APIKey)
-	//-- Output
-	hornbillHelpers.Logger(2, "---- XMLMC Database Asset Relationship Import Utility V"+version+" ----", true, logFileName)
-	hornbillHelpers.Logger(2, "Flag - Config File "+configFileName, true, logFileName)
 
-	configManager, err = isConfigManagerInstalled()
-	if err != nil {
-		hornbillHelpers.Logger(4, "Error checking for Configuration Manager installation: "+err.Error(), true, logFileName)
-	}
-
-	if !configManager {
-		hornbillHelpers.Logger(5, "Configuration Manager is not installed. Basic asset links ONLY will be imported", true, logFileName)
-	} else {
-		hornbillHelpers.Logger(2, "Configuration Manager has been detected. Basic asset links AND Configuration Manager dependencies/impacts will be imported", true, logFileName)
-	}
+	checkVersion()
+	logger(2, "---- XMLMC Database Asset Relationship Import Utility V"+version+" ----", true, true)
+	logger(2, "Flag - Config File "+configFileName, true, true)
 
 	cacheHornbillRecords()
 
@@ -72,7 +60,7 @@ func main() {
 	}
 
 	if len(assetRelationships) == 0 && len(assetDeleteRelationships) == 0 {
-		hornbillHelpers.Logger(4, "No asset relationship or removal records returned from database queries", true, logFileName)
+		logger(4, "No asset relationship or removal records returned from database queries", true, true)
 		os.Exit(1)
 	}
 
@@ -85,32 +73,32 @@ func main() {
 	}
 
 	//Output
-	hornbillHelpers.Logger(2, "Processing Complete!", true, logFileName)
-	hornbillHelpers.Logger(2, "* Relationship Records Found: "+strconv.Itoa(len(assetRelationships)), true, logFileName)
-	hornbillHelpers.Logger(2, "* Asset Links Created: "+strconv.Itoa(counters.linksCreated), true, logFileName)
-	hornbillHelpers.Logger(2, "* Asset Links Skipped (already exists): "+strconv.Itoa(counters.linksSkipped), true, logFileName)
-	hornbillHelpers.Logger(2, "* Asset Links Failed: "+strconv.Itoa(counters.linksFailed), true, logFileName)
-	hornbillHelpers.Logger(2, "* Dependency Records Created: "+strconv.Itoa(counters.depsCreated), true, logFileName)
-	hornbillHelpers.Logger(2, "* Dependency Records Updated: "+strconv.Itoa(counters.depsUpdated), true, logFileName)
-	hornbillHelpers.Logger(2, "* Dependency Records Skipped: "+strconv.Itoa(counters.depsSkipped), true, logFileName)
-	hornbillHelpers.Logger(2, "* Dependency Records Failed: "+strconv.Itoa(counters.depsFailed), true, logFileName)
-	hornbillHelpers.Logger(2, "* Dependency Records Update Failed: "+strconv.Itoa(counters.depsUpdateFailed), true, logFileName)
-	hornbillHelpers.Logger(2, "* Impact Records Created: "+strconv.Itoa(counters.impsCreated), true, logFileName)
-	hornbillHelpers.Logger(2, "* Impact Records Updated: "+strconv.Itoa(counters.impsUpdated), true, logFileName)
-	hornbillHelpers.Logger(2, "* Impact Records Skipped: "+strconv.Itoa(counters.impsSkipped), true, logFileName)
-	hornbillHelpers.Logger(2, "* Impact Records Failed: "+strconv.Itoa(counters.impsFailed), true, logFileName)
-	hornbillHelpers.Logger(2, "* Impact Records Update Failed: "+strconv.Itoa(counters.impsUpdateFailed), true, logFileName)
+	logger(2, "Processing Complete!", true, true)
+	logger(2, "* Relationship Records Found: "+strconv.Itoa(len(assetRelationships)), true, true)
+	logger(2, "* Asset Links Created: "+strconv.Itoa(counters.linksCreated), true, true)
+	logger(2, "* Asset Links Skipped (already exists): "+strconv.Itoa(counters.linksSkipped), true, true)
+	logger(2, "* Asset Links Failed: "+strconv.Itoa(counters.linksFailed), true, true)
+	logger(2, "* Dependency Records Created: "+strconv.Itoa(counters.depsCreated), true, true)
+	logger(2, "* Dependency Records Updated: "+strconv.Itoa(counters.depsUpdated), true, true)
+	logger(2, "* Dependency Records Skipped: "+strconv.Itoa(counters.depsSkipped), true, true)
+	logger(2, "* Dependency Records Failed: "+strconv.Itoa(counters.depsFailed), true, true)
+	logger(2, "* Dependency Records Update Failed: "+strconv.Itoa(counters.depsUpdateFailed), true, true)
+	logger(2, "* Impact Records Created: "+strconv.Itoa(counters.impsCreated), true, true)
+	logger(2, "* Impact Records Updated: "+strconv.Itoa(counters.impsUpdated), true, true)
+	logger(2, "* Impact Records Skipped: "+strconv.Itoa(counters.impsSkipped), true, true)
+	logger(2, "* Impact Records Failed: "+strconv.Itoa(counters.impsFailed), true, true)
+	logger(2, "* Impact Records Update Failed: "+strconv.Itoa(counters.impsUpdateFailed), true, true)
 	if importConf.RemoveLinks {
-		hornbillHelpers.Logger(2, "* Remove Relationship Records Found: "+strconv.Itoa(len(assetDeleteRelationships)), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Asset Links Success: "+strconv.Itoa(counters.removeLinksSuccess), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Asset Links Skipped (doesn't exist): "+strconv.Itoa(counters.removeLinksSkipped), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Asset Links Failed: "+strconv.Itoa(counters.removeLinksFailed), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Dependency Records Success: "+strconv.Itoa(counters.removeDepsSuccess), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Dependency Records Skipped: "+strconv.Itoa(counters.removeDepsSkipped), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Dependency Records Failed: "+strconv.Itoa(counters.removeDepsFailed), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Impact Records Success: "+strconv.Itoa(counters.removeImpsSuccess), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Impact Records Skipped: "+strconv.Itoa(counters.removeImpsSkipped), true, logFileName)
-		hornbillHelpers.Logger(2, "* Remove Impact Records Failed: "+strconv.Itoa(counters.removeImpsFailed), true, logFileName)
+		logger(2, "* Remove Relationship Records Found: "+strconv.Itoa(len(assetDeleteRelationships)), true, true)
+		logger(2, "* Remove Asset Links Success: "+strconv.Itoa(counters.removeLinksSuccess), true, true)
+		logger(2, "* Remove Asset Links Skipped (doesn't exist): "+strconv.Itoa(counters.removeLinksSkipped), true, true)
+		logger(2, "* Remove Asset Links Failed: "+strconv.Itoa(counters.removeLinksFailed), true, true)
+		logger(2, "* Remove Dependency Records Success: "+strconv.Itoa(counters.removeDepsSuccess), true, true)
+		logger(2, "* Remove Dependency Records Skipped: "+strconv.Itoa(counters.removeDepsSkipped), true, true)
+		logger(2, "* Remove Dependency Records Failed: "+strconv.Itoa(counters.removeDepsFailed), true, true)
+		logger(2, "* Remove Impact Records Success: "+strconv.Itoa(counters.removeImpsSuccess), true, true)
+		logger(2, "* Remove Impact Records Skipped: "+strconv.Itoa(counters.removeImpsSkipped), true, true)
+		logger(2, "* Remove Impact Records Failed: "+strconv.Itoa(counters.removeImpsFailed), true, true)
 	}
 }
 
@@ -119,32 +107,30 @@ func cacheHornbillRecords() {
 	//-- Cache Assets first
 	err := cacheAssets()
 	if err != nil {
-		hornbillHelpers.Logger(4, "Error when caching assets from Hornbill: "+err.Error(), true, logFileName)
+		logger(4, "Error when caching assets from Hornbill: "+err.Error(), true, true)
 		os.Exit(1)
 	}
 
 	//--Cache Links
 	err = cacheAssetLinks()
 	if err != nil {
-		hornbillHelpers.Logger(4, "Error when caching asset links from Hornbill: "+err.Error(), true, logFileName)
+		logger(4, "Error when caching asset links from Hornbill: "+err.Error(), true, true)
 		os.Exit(1)
 	}
 
-	if configManager {
-		//Cache Config Manager Asset Records
-		//--Cache Dependencies
-		err = cacheAssetDependencies()
-		if err != nil {
-			hornbillHelpers.Logger(4, "Error when caching asset dependencies from Hornbill: "+err.Error(), true, logFileName)
-			os.Exit(1)
-		}
+	//Cache Config Manager Asset Records
+	//--Cache Dependencies
+	err = cacheAssetDependencies()
+	if err != nil {
+		logger(4, "Error when caching asset dependencies from Hornbill: "+err.Error(), true, true)
+		os.Exit(1)
+	}
 
-		//--Cache Impact Records
-		err = cacheAssetImpacts()
-		if err != nil {
-			hornbillHelpers.Logger(4, "Error when caching asset impacts from Hornbill: "+err.Error(), true, logFileName)
-			os.Exit(1)
-		}
+	//--Cache Impact Records
+	err = cacheAssetImpacts()
+	if err != nil {
+		logger(4, "Error when caching asset impacts from Hornbill: "+err.Error(), true, true)
+		os.Exit(1)
 	}
 }
 
@@ -153,16 +139,16 @@ func loadConfig() sqlImportConfStruct {
 	//-- Check Config File File Exists
 	cwd, _ := os.Getwd()
 	configurationFilePath := cwd + "/" + configFileName
-	hornbillHelpers.Logger(1, "Loading Config File: "+configurationFilePath, false, logFileName)
+	logger(1, "Loading Config File: "+configurationFilePath, false, false)
 	if _, fileCheckErr := os.Stat(configurationFilePath); os.IsNotExist(fileCheckErr) {
-		hornbillHelpers.Logger(4, "No Configuration File", true, logFileName)
+		logger(4, "No Configuration File", true, true)
 		os.Exit(102)
 	}
 	//-- Load Config File
 	file, fileError := os.Open(configurationFilePath)
 	//-- Check For Error Reading File
 	if fileError != nil {
-		hornbillHelpers.Logger(4, "Error Opening Configuration File: "+fmt.Sprintf("%v", fileError), true, logFileName)
+		logger(4, "Error Opening Configuration File: "+fmt.Sprintf("%v", fileError), true, false)
 	}
 
 	//-- New Decoder
@@ -173,33 +159,58 @@ func loadConfig() sqlImportConfStruct {
 	err := decoder.Decode(&esqlConf)
 	//-- Error Checking
 	if err != nil {
-		hornbillHelpers.Logger(4, "Error Decoding Configuration File: "+fmt.Sprintf("%v", err), true, logFileName)
+		logger(4, "Error Decoding Configuration File: "+fmt.Sprintf("%v", err), true, false)
 	}
 	//-- Return New Congfig
 	return esqlConf
 }
 
-func isConfigManagerInstalled() (bool, error) {
-	xmlApps, err := espXmlmc.Invoke("session", "getApplicationList")
-	if err != nil {
-		hornbillHelpers.Logger(4, err.Error(), true, logFileName)
-		return false, err
+//-- Check Latest
+func checkVersion() {
+	githubTag := &latest.GithubTag{
+		Owner:      "hornbill",
+		Repository: appName,
 	}
 
-	var xmlResponse methodCallResult
-	err = xml.Unmarshal([]byte(xmlApps), &xmlResponse)
+	res, err := latest.Check(githubTag, version)
 	if err != nil {
-		hornbillHelpers.Logger(4, err.Error(), true, logFileName)
-		return false, err
+		msg := "Unable to check utility version against Github repository: " + err.Error()
+		logger(4, msg, true, true)
+		return
 	}
-	if xmlResponse.Status != "ok" {
-		hornbillHelpers.Logger(4, xmlResponse.State.ErrorRet, true, logFileName)
-		return false, errors.New(xmlResponse.State.ErrorRet)
+	if res.Outdated {
+		msg := "v" + version + " is not latest, you should upgrade to " + res.Current + " by downloading the latest package from: https://github.com/hornbill/" + appName + "/releases/tag/v" + res.Current
+		logger(5, msg, true, true)
 	}
-	for _, v := range xmlResponse.Params.Apps {
-		if v.Name == "com.hornbill.configurationmanager" && v.Status == "installed" {
-			return true, nil
-		}
+}
+
+// espLogger -- Log to ESP
+func espLogger(message string, severity string) {
+	if configDryrun {
+		message = "[DRYRUN] " + message
 	}
-	return false, err
+	espXmlmc.SetParam("fileName", appName)
+	espXmlmc.SetParam("group", "general")
+	espXmlmc.SetParam("severity", severity)
+	espXmlmc.SetParam("message", message)
+	espXmlmc.Invoke("system", "logMessage")
+}
+
+func logger(t int, s string, outputToCLI, outputToESP bool) {
+	//-- Create Log Entry
+	var espLogType string
+	switch t {
+	case 1:
+		espLogType = "debug"
+	case 4:
+		espLogType = "error"
+	case 5:
+		espLogType = "warn"
+	default:
+		espLogType = "notice"
+	}
+	if outputToESP {
+		espLogger(s, espLogType)
+	}
+	hornbillHelpers.Logger(t, s, outputToCLI, logFileName)
 }